refactor: 테마 추가/삭제 api를 AdminController로 이동
